Factor out generic functionality into tile-dataset
Rasters and point clouds will share much of the same functionality -
lots of tiles, stored in a user-recognisable format, stored in Git LFS.
This refactor moves the tile-dataset code that is not specifically
point-cloud related into tile/tile_dataset.py
